+++ lib/metro/commands/make/state_managed_widget.dart
@@ -29,14 +29,19 @@ class _MakeStateManagedWidgetCommand extends NyCustomCommand {
     final widgetName =
         requireArgument(result, message: 'A widget name is required');
 
+    MetroProjectFile projectFile = MetroService.createMetroProjectFile(
+        widgetName,
+        prefix: RegExp(r'(_?widget)'));
+
     String cleanWidgetName =
-        widgetName.snakeCase.replaceAll(RegExp(r'(_?widget)'), "");
+        projectFile.name.snakeCase.replaceAll(RegExp(r'(_?widget)'), "");
 
     ReCase classReCase = ReCase(cleanWidgetName);
 
     String stubStatefulWidget = widgetStateManagedStub(classReCase);
     await MetroService.makeStateManagedWidget(
         classReCase.snakeCase, stubStatefulWidget,
-        forceCreate: result.hasForceFlag);
+        forceCreate: result.hasForceFlag,
+        creationPath: projectFile.creationPath);
   }
 }